cmake: Kernel and app memory space update
Fixes: #8441

This commit updates the handling of memory location for libraries in
Zephyr during linking.

It introduces a generic zephyr_set_property function which can be
extended in future.

At start is supports the property:
- MEMORY_SPACE kernel
which can be used to control memory space location for libraries when
KConfig setting CONFIG_APPLICATION_MEMORY is enabled.

Signed-off-by: Torsten Rasmussen <torsten.rasmussen@nordicsemi.no>