cmake: ensure ZEPHYR_CURRENT_MODULE_DIR is using cmake native path style
This commit fixes an issue in windows where zephyr_modules.txt contains
a \ as path separator.
This causes issues later when using this variable.
find_path is used to ensure that the path will look correct in CMake
style and find_path avoids issues later if multiple path seperators are
used, such as 'path/////to///somewhere when striping.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>