linker: align __data_ram/rom_start/end linker symbol names
Cleanup and preparation commit for linker script generator.

This commit aligns the symbols for _data_ram/rom to other symbols and in
such a way they follow consistent pattern which allows for linker script
and scatter file generation.

The symbols are named according to the section name they describe.
Section name is `data`

The following symbols are aligned in this commit:
-  __data_ram_start  -> __data_start
-  __data_ram_end    -> __data_end
-  __data_rom_start  -> __data_load_start

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>