scripts: gen_handles.py: prepare for scatter file handling
The current gen_handles.py script uses linker defined symbols.
For ld and armlink the linker defined symbols can be retrieved from the
elf image.

However, armlink and ld uses different symbols.
With ld those can be named explicitly, but for armlink the linker
decides the names. When linking with armlink a steering file is used for
mapping of the Zephyr defined linker symbols, such as `__device_start`
to the armlink defined symbol.

However, this doesn't change the symbol names in the resulting elf file.

Therefore knowledge of the linker in use must be passed to
gen_handles.py so that the correct symbol can be used for locating
devices.

For ld, this is: __device_start
For armlink, this is: Image$$device$$Base

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>