arm-ds: INPUT for zephyr_linker_section_configure is now multiarg
INPUT argument on zephyr_linker_section_configure is now a multi arg.
This allows users to specify a list of input section matches in a
single call which further enables the possibility of surrounding a group
of input selectors with linker symbols.

Extended zephyr_linker_symbol to support memory alignment.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>