temphack: renamed .intList --> intList
armlink auto creates symbols for elfsections in the form of
Image$$<section>$$Base, Image$$<section>$$Limit, etc.

However, referring to a section in the form: Image$$.intList$$Base gives
troubles in C code, so therefore section names are cleaned so that
leading dots are removed.

This results in a new challenge when the build system refers to sections
and expects they are named with a leading dot, such as: .intList.

Thus, temporary adjust '.intList' to just 'intList'.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>