cmake: linker: add targets ld_target and scatter_target
Adding tow targets:
ld_target      : generate a ld linker script and print it to std out
scatter_target : generate a scatter file and print it to std out

This is a debug commit for facilitating and testing the linker script
generator.

This commit shall be dropped when PR is taken out of draft.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>