cmake: menuconfig: create a dedicated menuconfig project
Fixes: #9573

Create a dedicated CMake project for menuconfig and related targets.
This allows the Xephyr project to have menuconfig and related targets
available even if CMake later fails due to invalid .config settings or
other CMake errors.

The build files (Makefile / ninja) or copied from the menuconfig project
to the main project. This means that in the situation where CMake fails
and no build files are generated for the main project, then the
menuconfig targets will be available.

When CMake succeeds for the main project, the temporary menuconfig build
files are overwritten, and thus the projects build files, including
menuconfig are available to the user.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>