squash! cmake: run `gen_handles.py` on app_smem_unaligned
cmake: rework linker script generation and linker prebuilt stages

This commit reworks the linker script generation and linking stages in
order to better support fixed section location as required by #38836.

Today we have the following generated linker scripts and the elf output
depending on the system configuration:
- linker_app_smem_unaligned.cmd --> app_smem_unaligned_prebuilt.elf
- linker_zephyr_prebuilt.cmd    --> zephyr_prebuilt.elf
- linker.cmd                    --> zephyr.elf

as not all linker scripts may be created and as there is a need for the
possibility to move gen handles earlier then those stages has been
renamed into more generic names so that with this commit we have:
- linker_zephyr_pre0.cmd --> zephyr_pre0.elf
- linker_zephyr_pre1.cmd --> zephyr_pre1.elf
- linker.cmd             --> zephyr.elf

This also means that is the stage zephyr_pre1 is not needed, then build
can go from `zephyr_pre0i.elf` to `zephyr.elf`.

To keep the current behavior of generating the isr table of what was
`zephyr_prebuilt` stage the code block contolling isr generation has
been relocated to be located after app_smem and device handle
generation.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>