scripts: west command common runner multi image support [wip]
This commit is wip.

This commit extends the `do_run_common()` method to support multi image
runner file.

The multi image runner uses the following yaml format to specify the
build directory of each image in the multi image build:
> images:
> - name: <image-1>
> - build_dir: <build_dir-image-1>
> - name: <image-2>
> - build_dir: <build_dir-image-2>
> - name: ...
> - build_dir: ...

The `do_run_common()` has been extended so that the presence of
`multi_image_runners.yml` file in the build directory indicate a multi
image build syste, in which case the runner of each image is loaded and
executed, for example during `west flash`.

Still TODO:
- Support a `--image <name>` argument to specify only a specific runner
  in an image should be invoked, for example:
  `west debug --image mcuboot` to debug mcuboot.
  Or:
  `west flash --image blinky` to (re-)flash only blinky and not all
  images.
- Support a default runner for operations where only single runner
  should be invoked, for example `debug`.
  Default should be the application image

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>