scripts: west command common runner domain support
This commit extends the `do_run_common()` method to support domain
runner file and execution of multiple runner files in sequence.

The domain runner uses the following yaml format to specify the
build directory of each domain in the multi image build:
> default: <domain-n>
> domains:
>   <domain-1>:
>     build_dir: <build_dir-domain-1>
>   <domain-2>:
>     build_dir: <build_dir-domain-2>
>   ...

The `do_run_common()` has been extended so that the presence of
`domain_runners.yml` file in the build directory indicate a multi
image build system, in which case the runner of each domain is loaded
and executed, for example during `west flash`.

Default all domains are flashed, but `--domain <domain>` argument can
be used to select a specific domain to flash, for example:
> west flash --domain mcuboot

For west commands where only a single domain can be selected at a time,
for example `west debug`, then the default domain specified in the
domain_runners.yml file will be used.
Users can still select a different domain, for example with:
> west debug --domain mcuboot

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>