cmake: initial sysbuild / multi image support
This is the initial commit with sysbuild / Zephyr multi image build
support.

Using CMake as infrastructure together with the Zephyr sysbuild allows
us to support a convenient way of building a sample and allow user to
simply enable a bootloader and build that bootloader as part of the
sample build.

It support building any Zephyr sample together with mcuboot as a single
build invocation.

It uses Kconfig for configuration of image builds.
This allows for future extension with more bootloaders as well as
support for additional build images.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>