kconfig: support passing BOARD information from CMake to Kconfig
Today, board Kconfig files defines their own BOARD setting in
Kconfig.defconfig.

This is fragile because the BOARD used by the build system (CMake) can
easily become out of sync with the BOARD specified by Kconfig.

Furthermore, CONFIG_BOARD may not be equal to BOARD because of this,
and when testing in CMake one will use BOARD but when testing in C-code
then the CONFIG_BOARD value is used.

In a number of cases in Zephyr today, the BOARD and CONFIG_BOARD are
not matching.

This commit provides the possibility to use the CMake board for the
Kconfig setting so that they are always in sync, while still allow
the old style of definig BOARD in Kconfig.defconfig.

This means that Zephyr and and downstream users of Zephyr can transition
to the new and more robust style gradually.

To differentiate between old and new style, the new style will use
Kconfig.<board> file.

Specify board scheme v* for compliance and doc builds as this includes
both schemes into compliance and doc build.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>