boards: update nRF52840dk_nrf52811 board to board scheme v2.
This commit updates nrf52840dk_nrf52811 to use board scheme v2, and
thereby remove the Kconfig BOARD setting, as this is now passed from
CMake.

It also uses the new SoC approach, and thereby ensuring that the right
SoC is always used, and cannot accidentially be changed by the user.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>