libc/picolibc: Use libc-level PICOLIBC_MODULE to control picolibc mode
C++ is supported with picolibc only when the toolchain version of picolibc
is use -- libstdc++ must be built with picolibc support and libstdc++ is
included with the toolchain. That creates a circular dependency if the
control of whether to use the module is placed under the picolibc
configuration block (as PICOLIBC_USE_MODULE is).

Make an explicit PICOLIBC_MODULE variable at the libc level to be the
user-visible control for this feature and make code depending on whether
the picolibc module should be used depend on both PICOLIBC and
PICOLIBC_MODULE.

Change PICOLIBC_USE_MODULE to be a hidden configuration variable that the
picolibc module cmake files can use to determine whether to include the
picolibc module sources in the zephyr application build. This symbol can be
removed when the picolibc module is updated to use PICOLIBC AND
PICOLIBC_MODULE instead

Signed-off-by: Keith Packard <keithp@keithp.com>