soc: introduce list_socs.py for listing of SoCs
The list_socs.py script parses socs.yml in all <soc-root>/soc folders.

The socs.yml is introduced with hw model v2.

Hw model v2 removes the need for architecture knowledge of the SoCs,
and as part of this makes multi-arch and multi-core SoCs possible.

Hw model v2 also allows for greater flexibility in SoC organization as
SoCs can be organized freely.
As example SoCs can be organized by vendors, architecture, or any other
way as the socs.yml contains the path to the location of the SoC,
instead of relying on a specific arch.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>