scripts introduce list_hwm.py for listing of architectures and SoCs
The list_hwm.py script parses archs.yml in all <arch-root>/arch folders
and socs.yml in all <soc-root>/soc folders.

The archs.yml and socs.yml are introduced with hw model v2.

Hw model v2 removes the need for architecture knowledge of the SoCs,
and as part of this makes multi-arch and multi-core SoCs possible.

Hw model v2 also allows for greater flexibility in arch and SoC
organization as they can be organized freely.

As example SoCs can be organized by vendors, architecture, or any other
way as the socs.yml contains the path to the location of the SoC,
instead of relying on a specific arch.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>