cmake: extend zephyr string escape to handle double quote '"' char
A single escaped '\"' results in single '"' when written to disk.
The current escape functionality expands to '\\"', which works for chars
like '\' in '\n' to '\\n'.
With '\"', then '\\"' results in '"' to become string encapsulating.
Thus, escape must escape a '\"' to '\\\"' to ensure the first two '\\'
becomes a literal '\' and the '\"' becomes a literal '"'.

Result '\"' --> '\\\"'.
Example: 'foo: \"bar\'' --> 'foo: \\\"bar\\\"'

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>